convergence with master, take 5

+++ h2/src/main/org/h2/engine/Database.java
@@ -343,7 +343,7 @@ public class Database implements DataHandler {
         }
     }
 
-    public long getLockTimeout() {
+    public int getLockTimeout() {
         Setting setting = findSetting(
                 SetTypes.getTypeName(SetTypes.DEFAULT_LOCK_TIMEOUT));
         return setting == null ? Constants.INITIAL_LOCK_TIMEOUT : setting.getIntValue();
